Bug 797117: Explicitly cast UINT16_MAX to uint32_t in comparison within nsSocketTransportService::DoPollIteration(), to prevent it from being treated as signed and causing a build warning. r=ehsan
UltraBlame original commit: e474cc38930bc3165d75bdfd6bfaf690dfcecfbe

+++ netwerk/base/src/nsSocketTransportService2.cpp
@@ -790,7 +790,13 @@ nsSocketTransportService::DoPollIteration(bool wait)
             
             else if (s.mHandler->mPollTimeout != UINT16_MAX) {
                 
-                if (NS_UNLIKELY(pollInterval > (UINT16_MAX - s.mElapsedTime)))
+                
+                
+                
+                
+                if (NS_UNLIKELY(pollInterval >
+                                static_cast<uint32_t>(UINT16_MAX) -
+                                s.mElapsedTime))
                     s.mElapsedTime = UINT16_MAX;
                 else
                     s.mElapsedTime += uint16_t(pollInterval);
